Move display status function
