+++ src/oled_display.cpp
@@ -17,6 +17,8 @@
 #include <oled_display.h>
 #include <iohcCryptoHelpers.h>
 #include <iohcRemoteMap.h>
+#include <interact.h>
+#include <WiFi.h>
 
 Adafruit_SSD1306 display(SCREEN_WIDTH, SCREEN_HEIGHT, &Wire, OLED_RST);
 
@@ -60,5 +62,43 @@ void display1WAction(const uint8_t *remote, const char *action, const char *dir,
     }
     display.print("Action: ");
     display.println(action);
+    display.display();
+}
+
+void updateDisplayStatus() {
+    display.clearDisplay();
+    display.setCursor(0, 0);
+    display.print("WiFi: ");
+    switch (wifiStatus) {
+        case ConnState::Connected:
+            display.println("connected");
+            break;
+        case ConnState::Connecting:
+            display.println("connecting");
+            break;
+        default:
+            display.println("disconnected");
+            break;
+    }
+    display.setCursor(0, 10);
+    display.print("IP: ");
+    if (wifiStatus == ConnState::Connected) {
+        display.println(WiFi.localIP());
+    } else {
+        display.println("-");
+    }
+    display.setCursor(0, 20);
+    display.print("MQTT: ");
+    switch (mqttStatus) {
+        case ConnState::Connected:
+            display.println("connected");
+            break;
+        case ConnState::Connecting:
+            display.println("connecting");
+            break;
+        default:
+            display.println("disconnected");
+            break;
+    }
     display.display();
 }
